feat: enhance sync_meta script to manage branch protection rules for master
little unpleasant copy of lib source code because https://github.com/PyGithub/PyGithub/pull/2211 is still waiting

(disclaimer: as the last commit as of today was made on Dec 2021, i.e. 7 months ago, and there are lots of PRs open, i doubt it will be merged soon - is PyGithub still mainted, and by whom? anyways)

i originally wanted to make an equivalent of get_required_linear_history for setting it but the edit/PATCH action requires more fields (which edit_protections fills in if not given) so I went and made my own edit_protections as a whole so i have not copied partial implementation and its easier to refactor when PyGithub changes its edit_protection